requirements.txt
gql
requests
requests_toolbelt
django-crontab
celery
django-celery-beat
redis